A batchtools future is a future whose value will be resolved via batchtools

A batchtools future is a future whose value will be resolved via batchtools

Usage

BatchtoolsCustomFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsBashFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsFuture(
  expr = NULL,
  envir = parent.frame(),
  substitute = TRUE,
  globals = TRUE,
  packages = NULL,
  label = NULL,
  resources = list(),
  workers = NULL,
  finalize = getOption("future.finalize", TRUE),
  conf.file = findConfFile(),
  cluster.functions = NULL,
  registry = list(),
  ...
)

BatchtoolsMultiprocessFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsMulticoreFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsTemplateFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsLsfFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsOpenLavaFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsSGEFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsSlurmFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsTorqueFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsSSHFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  workers = availableCores(),
  ...
)

BatchtoolsUniprocessFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsLocalFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

BatchtoolsInteractiveFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  ...
)

Arguments

expr

The R expression to be evaluated

substitute

Controls whether expr should be substitute():d or not.

envir

The environment in which global environment should be located.

globals

(optional) a logical, a character vector, a named list, or a Globals object. If TRUE, globals are identified by code inspection based on expr and tweak searching from environment envir. If FALSE, no globals are used. If a character vector, then globals are identified by lookup based their names globals searching from environment envir. If a named list or a Globals object, the globals are used as is.

label

(optional) Label of the future (where applicable, becomes the job name for most job schedulers).

resources

(optional) A named list passed to the batchtools template (available as variable resources). See Section 'Resources' in batchtools::submitJobs() more details.

workers

(optional) The maximum number of workers the batchtools backend may use at any time. Interactive and "local" backends can only process one future at the time (workers = 1L), whereas HPC backends, where futures are resolved via separate jobs on a scheduler, can have multiple workers. In the latter, the default is workers = NULL, which will resolve to getOption("future.batchtools.workers"). If neither are specified, then the default is 100.

finalize

If TRUE, any underlying registries are deleted when this object is garbage collected, otherwise not.

conf.file

(optional) A batchtools configuration file.

cluster.functions

(optional) A batchtools ClusterFunctions object.

registry

(optional) A named list of settings to control the setup of the batchtools registry.

...

Additional arguments passed to future::Future().

Value

A BatchtoolsFuture object