A batchtools backend that resolves futures in parallel via background R sessions over SSH

A batchtools backend that resolves futures in parallel via background R sessions over SSH

Usage

batchtools_ssh(..., workers = availableWorkers(), fs.latency = 65)

Arguments

workers

The number of multicore processes to be available for concurrent batchtools multicore futures.

fs.latency

[numeric(1)]
Expected maximum latency of the file system, in seconds. Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to access files and directories. Usually safe to set to 0 to disable the heuristic, e.g. if you are working on a local file system.

Details

The batchtools_ssh backend uses the batchtools backend set up by batchtools::makeClusterFunctionsSSH(), which requires system commands ssh and ps as available on Linux and macOS.

An alternative to batchtools_ssh is to use cluster futures of the future package with a single local background session, i.e. plan(cluster, workers = c("localhost")).