A batchtools LSF backend resolves futures in parallel via a Load Sharing Facility (LSF) job scheduler

A batchtools LSF backend resolves futures in parallel via a Load Sharing Facility (LSF) job scheduler

Usage

batchtools_lsf(
  ...,
  template = "lsf",
  scheduler.latency = 1,
  fs.latency = 65,
  resources = list(),
  workers = getOption("future.batchtools.workers", default = 100L)
)

Arguments

template

(optional) Name of job-script template to be searched for by batchtools::findTemplateFile(). If not found, it defaults to the templates/lsf.tmpl part of this package (see below).

scheduler.latency

[numeric(1)]
Time to sleep after important interactions with the scheduler to ensure a sane state. Currently only triggered after calling submitJobs.

fs.latency

[numeric(1)]
Expected maximum latency of the file system, in seconds. Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to access files and directories. Usually safe to set to 0 to disable the heuristic, e.g. if you are working on a local file system.

workers

The maximum number of workers the batchtools backend may use at any time, which for HPC schedulers corresponds to the maximum number of queued jobs. The default is getOption("future.batchtools.workers", 100).

...

Not used.

Details

Batchtools lsf futures use batchtools cluster functions created by batchtools::makeClusterFunctionsLSF(), which requires that LSF commands bsub, bjobs, and bkill are installed on the current machine.

The default template script templates/lsf.tmpl can be found in:

system.file("templates", "lsf.tmpl", package = "future.batchtools")

and comprise:

#!/bin/bash
######################################################################
# A batchtools launch script template for LSF
#
# Author: Henrik Bengtsson
######################################################################

## Job name:
#BSUB -J <%= job.name %>

## Direct streams to logfile:
#BSUB -o <%= log.file %>

## Resources needed:
<% if (length(resources) > 0) {
  ## As-is resource specifications
  if ("asis" %in% names(resources)) {
    lines <- resources[["asis"]]
    resources <- resources[names(resources) != "asis"]
  } else {
    lines <- character(0)
  }

  ## Remaining resources are assumed to be of type '-<key>=<value>'
  opts <- unlist(resources, use.names = TRUE)
  opts <- sprintf("-%s=%s", names(opts), opts)
  lines <- sprintf("#BSUB %s", c(lines, opts))
  writeLines(lines)
} %>

echo "Batchtools information:"
echo "- job name: '<%= job.name %>'"
echo "- job log file: '<%= log.file %>'"
echo

echo "Session information:"
echo "- timestamp: $(date)"
echo "- hostname: $(hostname)"
echo "- Rscript path: $(which Rscript)"
echo "- Rscript version: $(Rscript --version)"
echo "- Rscript library paths: $(Rscript -e "cat(shQuote(.libPaths()), sep = ' ')")"
echo

echo "Job submission declarations:"
<%
  writeLines(sprintf("echo '%s'", lines))
%>

echo "Job information:"
bjobs -l "${LSB_JOBID}"
echo

## Launch R and evaluate the batchtools R job
echo "Command: Rscript -e 'batchtools::doJobCollection("<%= uri %>")' ..."
Rscript -e 'batchtools::doJobCollection("<%= uri %>")'
res=$?
echo " - exit code: ${res}"
echo "Command: Rscript -e 'batchtools::doJobCollection("<%= uri %>")' ... done"

## End-of-job summary
bjobs -l "${LSB_JOBID}"

## Relay the exit code from Rscript
exit "${res}"

References

https://en.wikipedia.org/wiki/IBM_Spectrum_LSF

Examples

if (FALSE) { # interactive()
library(future)

# Limit runtime to 10 minutes and total memory to 400 MiB per future,
# request a parallel environment with four slots on a single host.
# Submit to the 'freecycle' queue.
plan(future.batchtools::batchtools_lsf, resources = list(
  W = "00:10:00", M = "400",
  asis = c("-n 4", "-R 'span[hosts=1]'", "-q freecycle")
))

f <- future(Sys.info())
info <- value(f)
print(info)
}